+++ src/aux-common/rpc/ErrorCodes.ts
@@ -70,7 +70,8 @@ export type KnownErrorCodes =
     | 'not_found'
     | 'invalid_connection_state'
     | 'user_already_exists'
-    | 'session_is_not_revokable';
+    | 'session_is_not_revokable'
+    | 'hume_api_error';
 
 /**
  * Gets the status code that should be used for the given response.
@@ -176,6 +177,8 @@ export function getStatusCode(
             return 500;
         } else if (response.errorCode === 'user_already_exists') {
             return 400;
+        } else if (response.errorCode === 'hume_api_error') {
+            return 500;
         } else {
             return 400;
         }